requirements.txt
# Chatterbox Multilingual TTS Fine-Tuning Dependencies
# Tested with Python 3.11

# Core
torch>=2.1.0
torchaudio>=2.1.0

# Chatterbox TTS
chatterbox-tts

# HuggingFace datasets
datasets>=2.15.0
transformers>=4.36.0

# Model format conversion
safetensors>=0.4.0

# Visualization
matplotlib>=3.7.0

# Audio I/O
soundfile>=0.12.0

# Utilities
numpy>=1.24.0
tqdm>=4.65.0
